Validations added.

+++ BookManagement.Business/BookBusiness.cs
@@ -1,5 +1,8 @@
 ﻿using BookManagement.Model;
+using BookManagement.Model.Exceptions;
 using BookManagement.Model.Interfaces;
+using BookManagement.Model.Validators;
+using FluentValidation;
 
 namespace BookManagement.Business
 {
@@ -14,6 +17,11 @@ namespace BookManagement.Business
 
         public void Insert(Book book)
         {
+            var result = new BookValidator().Validate(book);
+
+            if (!result.IsValid)
+                throw new InvalidBookException($"Can't create new register: {string.Join("," , result.Errors)}");
+
             _bookRepository.Insert(book);
         }
 
@@ -24,16 +32,33 @@ namespace BookManagement.Business
 
         public Book Get(Guid id)
         {
-            return _bookRepository.Get(id);
+            Book? book = _bookRepository.Get(id);
+
+            if (book is null)
+                throw new InvalidBookException($"There is not a book with the id: {id}. Try again.");
+
+            else return book;
+            
         }
 
         public void Delete(Guid id)
         {
+            if (id.Equals(Guid.Empty))
+                throw new InvalidBookException("Please, inform a valid Id.");
+
+            if(!_bookRepository.CheckIfInserted(id))
+                throw new InvalidBookException($"The book with id {id} was not found. Please, try again.");
+
             _bookRepository.Delete(id);
         }
 
         public void Update(Book book)
         {
+            if(book.Id.Equals(Guid.Empty) 
+                || !_bookRepository.CheckIfInserted(book.Id)
+                || !new BookValidator().Validate(book).IsValid)
+                throw new InvalidBookException("Please, inform a valid book.");
+
             _bookRepository.Update(book);
         }
     }
